Rename sequential-only classes
